unittest improvements and new enpoint unit tests
