+++ tests/old/test_api_GET_methods.py
@@ -407,108 +407,7 @@ class GetSingleStudyTests(WsTests):
             self.assertEqual('NOT FOUND', err.reason)
 
 
-class GetStudyTitleTests(WsTests):
 
-    # GET Study Title - Pub -> 200
-    def test_get_title_pub(self):
-        request = urllib.request.Request(url_pub_id + '/title', method='GET')
-        with urllib.request.urlopen(request) as response:
-            self.assertEqual(response.code, 200)
-            header = response.info()
-            self.check_header_common(header)
-            body = response.read().decode('utf-8')
-            self.check_body_common(body)
-            j_resp = json.loads(body)
-            self.assertIsNotNone(j_resp['title'])
-
-    # GET Study Title - Pub - Auth -> 200
-    def test_get_title_pub_auth(self):
-        request = urllib.request.Request(url_pub_id + '/title', method='GET')
-        request.add_header('user_token', auth_id)
-        with urllib.request.urlopen(request) as response:
-            self.assertEqual(response.code, 200)
-            header = response.info()
-            self.check_header_common(header)
-            body = response.read().decode('utf-8')
-            self.check_body_common(body)
-            j_resp = json.loads(body)
-            self.assertIsNotNone(j_resp['title'])
-
-    # GET Study Title - Pub - NoAuth -> 200
-    def test_get_title_pub_noAuth(self):
-        request = urllib.request.Request(url_pub_id + '/title', method='GET')
-        request.add_header('user_token', wrong_auth_token)
-        with urllib.request.urlopen(request) as response:
-            self.assertEqual(response.code, 200)
-            header = response.info()
-            self.check_header_common(header)
-            body = response.read().decode('utf-8')
-            self.check_body_common(body)
-            j_resp = json.loads(body)
-            self.assertIsNotNone(j_resp['title'])
-
-    # GET Study Title - Priv - Auth -> 200
-    def test_get_title_priv_auth(self):
-        request = urllib.request.Request(url_priv_id + '/title', method='GET')
-        request.add_header('user_token', auth_id)
-        with urllib.request.urlopen(request) as response:
-            self.assertEqual(response.code, 200)
-            header = response.info()
-            self.check_header_common(header)
-            body = response.read().decode('utf-8')
-            self.check_body_common(body)
-            j_resp = json.loads(body)
-            self.assertIsNotNone(j_resp['title'])
-
-    # GET Study Title - Priv -> 403
-    def test_get_title_priv(self):
-        request = urllib.request.Request(url_priv_id + '/title', method='GET')
-        try:
-            urllib.request.urlopen(request)
-        except urllib.error.HTTPError as err:
-            self.assertEqual(err.code, 403)
-            self.check_header_common(err.headers)
-            self.check_body_common(err.read().decode('utf-8'))
-            self.assertEqual('FORBIDDEN', err.msg)
-            self.assertEqual('FORBIDDEN', err.reason)
-
-    # GET Study Title - Priv - NoAuth -> 403
-    def test_get_title_priv_noAuth(self):
-        request = urllib.request.Request(url_priv_id + '/title', method='GET')
-        request.add_header('user_token', wrong_auth_token)
-        try:
-            urllib.request.urlopen(request)
-        except urllib.error.HTTPError as err:
-            self.assertEqual(err.code, 403)
-            self.check_header_common(err.headers)
-            self.check_body_common(err.read().decode('utf-8'))
-            self.assertEqual('FORBIDDEN', err.msg)
-            self.assertEqual('FORBIDDEN', err.reason)
-
-    # GET Study Title - NullId -> 404
-    def test_get_title_nullId(self):
-        request = urllib.request.Request(url_null_id + '/title', method='GET')
-        try:
-            urllib.request.urlopen(request)
-        except urllib.error.HTTPError as err:
-            self.assertEqual(err.code, 404)
-            self.check_header_common(err.headers)
-            self.check_body_common(err.read().decode('utf-8'))
-            self.assertEqual('NOT FOUND', err.msg)
-            self.assertEqual('NOT FOUND', err.reason)
-
-    # GET Study Title - BadId -> 404
-    def test_get_title_badId(self):
-        request = urllib.request.Request(url_wrong_id + '/title', method='GET')
-        request.add_header('user_token', auth_id)
-        try:
-            urllib.request.urlopen(request)
-        except urllib.error.HTTPError as err:
-            self.assertEqual(err.code, 404)
-            self.check_header_common(err.headers)
-            self.check_body_common(err.read().decode('utf-8'))
-            self.assertEqual('NOT FOUND', err.msg)
-            self.assertEqual('NOT FOUND', err.reason)
 
 
 class GetStudyDescriptionTests(WsTests):
